Skip PUT when issue has not changed
Import deepdiff and use it to compare if the two (normalized)
issue dictionaries are equal. If they are, don't bother with a PUT.

The comparison is slightly complicated by two factors. First, the issue
from the API will contain an 'id' field, that has to be removed from
consideration. Second, the list of environments is unsorted, which
breaks the comparison.

issues_equal() deals with both of these cases, and a test has been added
as well for both situations.

Signed-off-by: Dan Rue <dan.rue@linaro.org>