Drop .netrc support; use an env variable instead
netrc was problematic for two reasons.

First, due to the caveat listed in README (removed in this commit) which
causes authentication to not work if a netrc file is located in
~/.netrc.

Second, because when running from travis, it is a bit awkward and likely
insecure to have to write out a .netrc file to the filesystem

Instead, KISS and require QA_REPORTS_KNOWN_ISSUE_TOKEN to be set. This
has the downside of not supporting multiple URLs; that is now going to
have to be handled separately. This isn't much of an issue in practice,
yet, so it may be deferred.

Signed-off-by: Dan Rue <dan.rue@linaro.org>