Find and print known issues that need to be rm'd
After syncing known issues, go through all of the known issues that the
API knows about and print a warning for any known issues that are in the
API but not listed in the corresponding yaml file. These issues are
candidates for deletion. In the future, delete could be automated.

Signed-off-by: Dan Rue <dan.rue@linaro.org>