Fix a bug that wasn't applying all the matrix envs
There was a bug in some matrix_apply's where the subsequent sets of
environments+projects would overwrite the previous set. Added a test and
fixed the issue.

Also discovered an unexpected behavior when accidentally passing a
string into 'test_names', rather than a list. Added a test and an assert
to catch that case.

Signed-off-by: Dan Rue <dan.rue@linaro.org>